Add files via upload
Silver-alginate beads were prepared and used for drug loading studies. A specific amount of these beads was incubated with a drug solution to facilitate drug loading through electrostatic interactions, entrapment, and diffusion. After incubation, the beads were removed, and the remaining drug concentration in the solution was measured to estimate the loading efficiency. Subsequently, the drug-loaded beads were placed in phosphate buffer solutions of different pH levels and stirred. At regular intervals, samples were withdrawn and analyzed for drug concentration to observe the release profile over time. The data collected were used to analyze the drug release kinetics, fitting the results to models such as zero-order, first-order, Higuchi, and Korsmeyer-Peppas to understand the rate and mechanism of drug release from the beads. From the analyzed data it was found that the drug release kinetics followed the first order kinetics.